C language

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,7 +9,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int counter = 0;
+	size_t counter = 0;
 
 	while (h != 0)
 	{
@@ -20,7 +20,7 @@ size_t print_list(const list_t *h)
 		}
 		else
 		{
-			printf("[%d] %s\n", h->len, h->str);
+			printf("[%u] %s\n", h->len, h->str);
 		}
 		h = h->next;
 	}
